requirements.txt
# requirements.txt (producción)
joblib>=1.4,<1.5
scipy>=1.15.2
streamlit>=1.39,<1.40

# Análisis de datos y machine learning
numpy==2.1.1
pandas==2.2.3
scikit-learn==1.5.2
optuna==3.6.1
xgboost==2.1.1
lightgbm==4.5.0

# Visualización
matplotlib==3.9.2
seaborn==0.13.2
plotly==5.24.1

# Jupyter y notebooks
jupyter==1.1.1
ipykernel==6.29.5
notebook==7.2.2

# Aplicación web
streamlit==1.39.0
gradio==4.44.1

# Base de datos
sqlalchemy==2.0.36
psycopg2-binary==2.9.10   # driver PostgreSQL

# Utilidades
tqdm==4.66.5
joblib==1.4.2
